move src/pageComponents/buy-me-coffee to app/buy-me-coffee/_components
