+++ template/web/app/buy-me-coffee/_components/ContractDemo.test.tsx
