+++ template/web/app/buy-me-coffee/_components/ContractDemo.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useMemo, useState } from 'react';
 import { clsx } from 'clsx';
 import useOnchainCoffeeMemos from '@/hooks/useOnchainCoffeeMemos';
