+++ template/web/app/buy-me-coffee/_components/steps/StartTransactionStep.tsx
