+++ template/web/app/buy-me-coffee/index.tsx
@@ -4,9 +4,9 @@ import { useEffect, useState } from 'react';
 import Banner from '@/components/banner/banner';
 import Header from '@/components/header/Header';
 import HorizontalLine from '@/components/horizontal-line/HorizontalLine';
-import BuyMeCoffeeContractDemo from '@/pageComponents/buy-me-coffee/ContractDemo';
-import Guide from '@/pageComponents/buy-me-coffee/Guide';
-import Profile from '@/pageComponents/buy-me-coffee/Profile';
+import BuyMeCoffeeContractDemo from './_components/ContractDemo';
+import Guide from './_components/Guide';
+import Profile from './_components/Profile';
 
 /**
  * Use the page component to wrap the components
